added delete(), get() and tests
